feat: [api-handlers-decomposition-002] - Preserve request-body decode and canonical work-request validation
