+++ docs/internal/processes/development-guide-relevant-files.md
@@ -136,7 +136,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `.github/workflows/ci.yml` | Repository CI contract | Contributor docs should name this workflow and mirror its root-level validation commands and stated non-deployment scope; when a required lane needs actionable diagnostics, prefer separate workflow steps that each call the existing repo-owned root commands such as `make ui-build`, `make build`, `make lint`, `make api-smoke`, `make ui-test`, `make ui-integration-test`, and `make test-coverage-go` instead of hiding the whole lane behind one opaque composite step. Keep Playwright Chromium installation immediately before the browser-backed `make ui-integration-test` step so CI and local review flows exercise the same dedicated lane. The backend coverage job should continue to call `make test-coverage-go` so CI and local runs share the same repo-owned `cmd/gocoveragecheck` policy surface. |
 | `internal/testpath/artifact_contract.go` | Enforced root artifact contract | Root-level factory artifact additions, removals, and redirect stubs are test-enforced here and must stay synchronized with the inventory doc. |
 | `factory/` | Maintainer workflow surface | Contains checked-in operator guidance and active inbox directories that the development guide may reference for workflow-related tasks. |
-| `pkg/api/handlers.go` plus sibling `pkg/api/handlers_*.go` files | Handwritten API boundary | Keep HTTP handler entrypoints, raw-JSON decode or validation helpers, response projection helpers, and shared API response writers in the same package but split across sibling `handlers_*.go` files when size cleanup is needed; shared JSON/SSE writers and `errorFamilyForStatus` live in `handlers_common.go`. The root `pkg/api` package has a 15-file cap—when adding a new `handlers_*.go` owner file, consolidate or move an existing root test/helper file first (for example colocate submit staged-file handlers with `submit_work_items.go`). Generated request structs still do not reject unknown keys by themselves, so retired work-request alias and trace-conflict policy should continue routing through `pkg/factory.ValidateCanonicalWorkRequestJSON(...)` and only be translated into API-owned error wording rather than reimplemented locally. |
+| `pkg/api/handlers.go` plus sibling `pkg/api/handlers_*.go` files | Handwritten API boundary | Keep HTTP handler entrypoints, raw-JSON decode or validation helpers, response projection helpers, and shared API response writers in the same package but split across sibling `handlers_*.go` files when size cleanup is needed; shared JSON/SSE writers and `errorFamilyForStatus` live in `handlers_common.go`; request-body decode helpers and canonical work-request validation translation live in `handlers_decode.go`. The root `pkg/api` package has a 15-file cap—when adding a new `handlers_*.go` owner file, consolidate or move an existing root source/test file first (for example colocate submit staged-file handlers with `submit_work_items.go`, or move `workstation_request_projection.go` into `pkg/api/workstationprojection/` with a thin `BuildFactoryWorldWorkstationRequestProjectionSlice` delegate in `handlers.go`). Generated request structs still do not reject unknown keys by themselves, so retired work-request alias and trace-conflict policy should continue routing through `pkg/factory/requests.ValidateCanonicalWorkRequestJSON(...)` via `validateCanonicalWorkRequestJSONForAPI` in `handlers_decode.go` and only be translated into API-owned error wording rather than reimplemented locally. |
 | `pkg/api/server_*_test.go` plus `pkg/api/server_test_helpers_test.go` | API route regression ownership | When API route tests cross the backend size gate, split them into same-package sibling `server_*_test.go` files by boundary concern such as submit, upsert, factory, provider-session, work-query, and dashboard/event coverage, and keep shared request builders or response assertions in `server_test_helpers_test.go` so route behavior stays observable without recreating one giant test owner file. |
 | `api/codegen_config/client.yaml` | Generated Go client config | The checked-in Go client for backend functional tests is generated from this config and must stay in lockstep with `api/openapi.yaml`, `pkg/generatedclient/`, and `make api-smoke`. |
 | `api/openapi-main.yaml` | Authored OpenAPI source of truth | Public API behavior changes should start here, including the advertised local `servers` URL, then refresh the bundled spec and generated Go/UI artifacts together so handwritten route seams and generated parameter binding stay aligned. |
