+++ pkg/api/workstationprojection/projection.go
@@ -1,4 +1,4 @@
-package api
+package workstationprojection
 
 import (
 	"sort"
