+++ pkg/api/workstationprojection/projection_content_assert_test.go
@@ -1,0 +1,128 @@
+package workstationprojection
+
+import (
+	"encoding/json"
+	"testing"
+
+	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
+	"github.com/portpowered/infinite-you/pkg/interfaces"
+)
+
+func assertGeneratedWorkContentParts(t *testing.T, content *factoryapi.WorkContent, want []interfaces.WorkContentPart) {
+	t.Helper()
+	if content == nil {
+		t.Fatalf("content = nil, want %#v", want)
+	}
+	if len(*content) != len(want) {
+		t.Fatalf("content count = %d, want %d", len(*content), len(want))
+	}
+	for i, wantPart := range want {
+		assertGeneratedWorkContentPart(t, (*content)[i], i, wantPart)
+	}
+}
+
+func assertGeneratedWorkContentPart(t *testing.T, got factoryapi.WorkContentPart, index int, want interfaces.WorkContentPart) {
+	t.Helper()
+	switch want.Type {
+	case interfaces.WorkContentPartTypeText:
+		assertGeneratedTextContentPart(t, got, index, want)
+	case interfaces.WorkContentPartTypeImage:
+		assertGeneratedImageContentPart(t, got, index, want)
+	case interfaces.WorkContentPartTypeAudio:
+		assertGeneratedAudioContentPart(t, got, index, want)
+	case interfaces.WorkContentPartTypeJSON:
+		assertGeneratedJSONContentPart(t, got, index, want)
+	case interfaces.WorkContentPartTypeBinary:
+		assertGeneratedBinaryContentPart(t, got, index, want)
+	default:
+		t.Fatalf("unsupported expected content type %q", want.Type)
+	}
+}
+
+func assertGeneratedTextContentPart(t *testing.T, got factoryapi.WorkContentPart, index int, want interfaces.WorkContentPart) {
+	t.Helper()
+	part, err := got.AsWorkTextContentPart()
+	if err != nil {
+		t.Fatalf("content[%d] decode text: %v", index, err)
+	}
+	if part.Type != factoryapi.WorkContentPartTypeText || part.Text != want.Text {
+		t.Fatalf("content[%d] = %#v, want text %q", index, part, want.Text)
+	}
+	assertGeneratedPartSharedFields(t, index, part.Slot, part.Label, part.Role, part.ContentType, part.ArtifactId, part.Metadata, want)
+}
+
+func assertGeneratedImageContentPart(t *testing.T, got factoryapi.WorkContentPart, index int, want interfaces.WorkContentPart) {
+	t.Helper()
+	part, err := got.AsWorkImageContentPart()
+	if err != nil {
+		t.Fatalf("content[%d] decode image: %v", index, err)
+	}
+	if part.Type != factoryapi.WorkContentPartTypeImage || part.File != want.File {
+		t.Fatalf("content[%d] = %#v, want image %q", index, part, want.File)
+	}
+	assertGeneratedPartSharedFields(t, index, part.Slot, part.Label, part.Role, part.ContentType, part.ArtifactId, part.Metadata, want)
+}
+
+func assertGeneratedAudioContentPart(t *testing.T, got factoryapi.WorkContentPart, index int, want interfaces.WorkContentPart) {
+	t.Helper()
+	part, err := got.AsWorkAudioContentPart()
+	if err != nil {
+		t.Fatalf("content[%d] decode audio: %v", index, err)
+	}
+	if part.Type != factoryapi.WorkContentPartTypeAudio || part.File != want.File {
+		t.Fatalf("content[%d] = %#v, want audio %q", index, part, want.File)
+	}
+	assertGeneratedPartSharedFields(t, index, part.Slot, part.Label, part.Role, part.ContentType, part.ArtifactId, part.Metadata, want)
+}
+
+func assertGeneratedJSONContentPart(t *testing.T, got factoryapi.WorkContentPart, index int, want interfaces.WorkContentPart) {
+	t.Helper()
+	part, err := got.AsWorkJsonContentPart()
+	if err != nil {
+		t.Fatalf("content[%d] decode json: %v", index, err)
+	}
+	rawJSON, err := json.Marshal(part.Json)
+	if err != nil {
+		t.Fatalf("content[%d] encode json: %v", index, err)
+	}
+	if part.Type != factoryapi.WorkContentPartTypeJSON || string(rawJSON) != string(want.JSON) {
+		t.Fatalf("content[%d] = %#v, want json %s", index, part, want.JSON)
+	}
+	assertGeneratedPartSharedFields(t, index, part.Slot, part.Label, part.Role, part.ContentType, part.ArtifactId, part.Metadata, want)
+}
+
+func assertGeneratedBinaryContentPart(t *testing.T, got factoryapi.WorkContentPart, index int, want interfaces.WorkContentPart) {
+	t.Helper()
+	part, err := got.AsWorkBinaryContentPart()
+	if err != nil {
+		t.Fatalf("content[%d] decode binary: %v", index, err)
+	}
+	if part.Type != factoryapi.WorkContentPartTypeBinary || part.File != want.File {
+		t.Fatalf("content[%d] = %#v, want binary %q", index, part, want.File)
+	}
+	assertGeneratedPartSharedFields(t, index, part.Slot, part.Label, part.Role, part.ContentType, part.ArtifactId, part.Metadata, want)
+}
+
+func assertGeneratedPartSharedFields(t *testing.T, index int, slot *string, label *string, role *string, contentType *string, artifactID *string, metadata *factoryapi.WorkContentMetadata, want interfaces.WorkContentPart) {
+	t.Helper()
+
+	if derefString(slot) != want.Slot ||
+		derefString(label) != want.Label ||
+		derefString(role) != want.Role ||
+		derefString(contentType) != want.ContentType ||
+		derefString(artifactID) != want.ArtifactID {
+		t.Fatalf("content[%d] shared fields mismatch for %#v", index, want)
+	}
+	gotMetadata, _ := json.Marshal(metadata)
+	wantMetadata, _ := json.Marshal(want.Metadata)
+	if string(gotMetadata) != string(wantMetadata) {
+		t.Fatalf("content[%d] metadata = %s, want %s", index, gotMetadata, wantMetadata)
+	}
+}
+
+func derefString(value *string) string {
+	if value == nil {
+		return ""
+	}
+	return *value
+}
