+++ pkg/api/workstationprojection/projection_test.go
@@ -1,4 +1,4 @@
-package api
+package workstationprojection
 
 import (
 	"bufio"
